Final stabilization pass: Fixed Safe Args navigation, optimized RecyclerView performance, cleaned up build artifacts, and updated README with showcase section.

+++ app/src/main/java/com/oceanx/freshcart/presentation/ordersuccess/OrderSuccessFragment.kt
@@ -47,7 +47,7 @@ class OrderSuccessFragment : Fragment(R.layout.fragment_order_success) {
      */
     private fun displayOrderDetails() {
         binding.orderIdValue.text = args.orderId
-        binding.totalAmountValue.text = args.totalAmount.toCurrencyString()
+        binding.totalAmountValue.text = args.totalAmount.toDouble().toCurrencyString()
         binding.deliveryAddressValue.text = args.deliveryAddress
         binding.paymentMethodValue.text = args.paymentMethod
     }
